Update implementation to use a single attribute converter
- Update the tests to rely on metadata only as previous image is no
  longer available
